\name{AdaptFit-internal}
\alias{aspm}
\alias{aspmFormRead}
\alias{aspmPenRead}
\alias{aspmKrigeRead}
\alias{aspGetTheta}
\alias{almeAux}
\alias{acompute.CTRinvC}
\alias{bdiag}
\alias{airls.ridge}
\title{Internal AdaptFit objects}
\description{Internal functions}
\details{These are not to be called by the user}
\keyword{internal}